music-ai/pyproject.toml
[tool.poetry]
name = "muai"
version = "0.1.0"
description = "A CLI tool for music vectorization and similarity search."
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.3.2"
librosa = "^0.10.0.post2"
numpy = "^1.21.1"
faiss-cpu = "^1.7.1"
soundfile = "^0.12.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
muai = "muai:app"

